Start implementing support for key commitment (incl. for the AD)
In order to do so, we output two AES blocks (256 bits) right after
initialization, prior to absorbing the AD.

This commitment is assumed to be send/stored/verified along with the
authentication tag. It is not a replacement for that tag. Both need
to be verified, hence it is reasonable for the commitment size to
match the key size and not twice its size.